ENH move things around a bit
